Fixed broken tests and added post and page comment tests
