+++ lib/webhookPublisher/publisher.ts
@@ -67,7 +67,6 @@ export async function publishWebhookEvent(spaceId: string, event: WebhookEvent) 
 
     // Add the message to the queue
     await addMessageToSQS(SQS_QUEUE_NAME, JSON.stringify(webhookPayload));
-    await createNotificationsFromEvent(webhookPayload);
     log.debug(`Sent webhook event to SQS: "${event.scope}"`, {
       spaceId,
       queueUrl: SQS_QUEUE_NAME,
